feat: support for DHT and without DHT at the same time refactor(tests): make tests independent without require global set up steps feat: move libp2p just to the network, take libp2p as a whole and not pieces docs: Update README Documentation
