+++ package.json
@@ -4,7 +4,8 @@
   "description": "Node.js implementation of the Bitswap data exchange protocol used by IPFS",
   "main": "src/index.js",
   "browser": {
-    "./test/libp2p-bundle": false
+    "./test/utils/create-libp2p-node": false,
+    "./test/utils/create-temp-repo-nodejs.js": "./test/utils/create-temp-repo-browser.js"
   },
   "scripts": {
     "test": "aegir-test",
@@ -39,10 +40,11 @@
   "devDependencies": {
     "aegir": "^11.0.2",
     "benchmark": "^2.1.4",
-    "chai": "^4.1.0",
+    "chai": "^4.1.1",
     "dirty-chai": "^2.0.1",
     "ipfs-repo": "~0.17.0",
     "libp2p": "^0.11.0",
+    "libp2p-kad-dht": "^0.4.1",
     "libp2p-multiplex": "^0.4.4",
     "libp2p-secio": "^0.7.1",
     "libp2p-tcp": "^0.10.2",
@@ -58,7 +60,7 @@
   "dependencies": {
     "async": "^2.5.0",
     "cids": "~0.5.1",
-    "debug": "^2.6.8",
+    "debug": "^3.0.0",
     "ipfs-block": "~0.6.0",
     "lodash.debounce": "^4.0.8",
     "lodash.find": "^4.6.0",
@@ -75,6 +77,7 @@
     "pull-length-prefixed": "^1.3.0",
     "pull-pushable": "^2.1.1",
     "pull-stream": "^3.6.0",
+    "safe-buffer": "^5.1.1",
     "varint-decoder": "^0.1.1"
   },
   "contributors": [
