+++ src/want-manager/index.js
@@ -2,10 +2,11 @@
 
 const debug = require('debug')
 
-const Message = require('../../types/message')
-const Wantlist = require('../../types/wantlist')
-const CONSTANTS = require('../../constants')
+const Message = require('../types/message')
+const Wantlist = require('../types/wantlist')
+const CONSTANTS = require('../constants')
 const MsgQueue = require('./msg-queue')
+const setImmediate = require('async/setImmediate')
 
 const log = debug('bitswap:wantmanager')
 log.error = debug('bitswap:wantmanager:error')
@@ -111,24 +112,24 @@ module.exports = class WantManager {
     this._stopPeerHandler(peerId)
   }
 
-  run () {
+  start (callback) {
+    // resend entire wantlist every so often
     this.timer = setInterval(() => {
-      // resend entirew wantlist every so often
       const fullwantlist = new Message(true)
-      for (let entry of this.wantlist.entries()) {
-        fullwantlist.addEntry(entry[1].cid, entry[1].priority)
-      }
-
-      this.peers.forEach((p) => {
-        p.addMessage(fullwantlist)
+      this.wantlist.forEach((entry) => {
+        fullwantlist.addEntry(entry.cid, entry.priority)
       })
+
+      this.peers.forEach((p) => p.addMessage(fullwantlist))
     }, 10 * 1000)
+
+    setImmediate(() => callback())
   }
 
-  stop () {
-    for (let mq of this.peers.values()) {
-      this.disconnected(mq.peerId)
-    }
+  stop (callback) {
+    this.peers.forEach((mq) => this.disconnected(mq.peerId))
+
     clearInterval(this.timer)
+    setImmediate(() => callback())
   }
 }
