+++ test/fixtures/serialized-from-go/bitswap110-message-full-wantlist
