+++ test/utils/create-libp2p-node.js
@@ -1,0 +1,42 @@
+'use strict'
+
+const TCP = require('libp2p-tcp')
+const Multiplex = require('libp2p-multiplex')
+const SECIO = require('libp2p-secio')
+const libp2p = require('libp2p')
+const KadDHT = require('libp2p-kad-dht')
+const waterfall = require('async/waterfall')
+const PeerInfo = require('peer-info')
+
+class Node extends libp2p {
+  constructor (peerInfo, options) {
+    options = options || {}
+
+    const modules = {
+      transport: [new TCP()],
+      connection: {
+        muxer: Multiplex,
+        crypto: SECIO
+      },
+      DHT: options.DHT ? KadDHT : undefined
+    }
+
+    super(modules, peerInfo, null, options.DHT || {})
+  }
+}
+
+function createLibp2pNode (options, callback) {
+  let node
+
+  waterfall([
+    (cb) => PeerInfo.create(cb),
+    (peerInfo, cb) => {
+      peerInfo.multiaddrs.add('/ip4/0.0.0.0/tcp/0')
+      node = new Node(peerInfo, options)
+      node.start(cb)
+    }
+  ], (err) => callback(err, node))
+}
+
+exports = module.exports = createLibp2pNode
+exports.bundle = Node
